Moved session, version and stats into the info cog
